squash_me

+++ contracts/beasts/src/constants.cairo
@@ -47,8 +47,8 @@ mod BeastSettings {
     // vs starter beasts which are always the same and maximally easy
     const STARTER_BEAST_LEVEL_THRESHOLD: u8 = 1;
 
-    const CRITICAL_HIT_LEVEL_MULTIPLIER: u8 = 1;
-    const CRITICAL_HIT_AMBUSH_MULTIPLIER: u8 = 3;
+    const CRITICAL_HIT_LEVEL_MULTIPLIER: u16 = 1;
+    const CRITICAL_HIT_AMBUSH_MULTIPLIER: u16 = 3;
 }
 
 mod BeastId {
